Refactor imports to use centralized icon path and utils path
