--- src/app/auth/create-admin/page.tsx
@@ -1,4 +1,0 @@
-export default function Home() {
-    return (<div>Create Admin</div>);
-  }
-
