+++ src/app/auth/sign-up/page.tsx
@@ -8,7 +8,6 @@ import FormInput from '@/components/FormInput';
 import LoadingButton from '@/components/LoadingButton';
 import NavLink from '@/components/NavLink';
 import PasswordValidation from '@/components/PasswordValidation';
-import BluewaveLogo from '../../../../public/assets/BluewaveLogo';
 import AuthFormWrapper from '../components/AuthFormWrapper';
 
 import { useFormSubmission, useValidatedFormData } from '@/hooks';
@@ -16,7 +15,8 @@ import {
 	passwordValidationRule,
 	requiredFieldRule,
 	validEmailRule,
-} from '@/utils/shared/validators';
+} from '@/shared/utils/validators';
+import BluewaveLogo from '@/components/layout/BluewaveLogo';
 
 export default function SignUp() {
 	const router = useRouter();
