+++ src/app/documents/[documentId]/components/InfoTableRow.tsx
@@ -3,7 +3,7 @@ import { useState } from 'react';
 
 import { Box, IconButton, TableCell, TableRow, Tooltip, Typography } from '@mui/material';
 
-import { CheckIcon, CopyIcon, TrashIcon } from '@/../public/assets/icons';
+import { CheckIcon, CopyIcon, TrashIcon } from '@/icons';
 
 import ModalWrapper from '@/components/ModalWrapper';
