+++ src/app/documents/components/SharingOptionsAccordion.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { Box, IconButton, MenuItem, RadioGroup, Select, Typography } from '@mui/material';
 
-import { EyeIcon, EyeOffIcon } from '@/../public/assets/icons';
+import { EyeIcon, EyeOffIcon } from '@/icons';
 
 import CustomCheckbox from '@/components/CustomCheckbox';
 import FormInput from '@/components/FormInput';
