+++ src/components/Breadcrumb.tsx
@@ -5,14 +5,7 @@ import { usePathname } from 'next/navigation';
 import { useMemo } from 'react';
 import NavLink from './NavLink';
 
-import {
-	ChevronRightIcon,
-	FileIcon,
-	HomeIcon,
-	SettingsIcon,
-	UserIcon,
-	UsersIcon,
-} from '@/../public/assets/icons';
+import { ChevronRightIcon, FileIcon, HomeIcon, SettingsIcon, UserIcon, UsersIcon } from '@/icons';
 
 const Breadcrumb = () => {
 	const pathname = usePathname();
