+++ src/providers/auth/AuthWrapper.tsx
@@ -1,5 +1,6 @@
 import SignIn from '@/app/auth/sign-in/page';
-import Sidebar from '@/app/layout/Sidebar';
+import Sidebar from '@/components/layout/Sidebar';
+
 import { Box, CircularProgress } from '@mui/material';
 import { useSession } from 'next-auth/react';
 import { usePathname } from 'next/navigation';
